Implement ConnectX-4 configurations
Add changes to implement features for ConnectX-4 adapters

+++ deployment_scripts/post_set_mellanox_interfaces.sh
@@ -21,22 +21,28 @@ source ./common
 if ([[ $ROLES == *compute* ]] && [[ ! $ROLES == "compute" ]]) \
     && [ $SRIOV == true ] ; then
 
-  # Update VFs
-  ./sriov.sh configure
+  if [ $CX == 'ConnectX-3' ]; then
 
-  # Kill tgt daemons if exists
-  tgt_locks=`find /var/run/ -name tgtd* | wc -l`
-  if [ $tgt_locks -ne 0 ];then
-    \rm -f /var/run/tgtd* && killall -9 tgtd
-    service tgt stop
+    # Update VFs
+    ./sriov.sh configure
+
+    # Kill tgt daemons if exists
+    tgt_locks=`find /var/run/ -name tgtd* | wc -l`
+    if [ $tgt_locks -ne 0 ];then
+      \rm -f /var/run/tgtd* && killall -9 tgtd
+      service tgt stop
+    fi
+
+    service openibd restart && service openvswitch-switch restart
+
+    if [ $tgt_locks -ne 0 ];then
+      service tgt start
+    fi
   fi
-
-  service openibd restart && service openvswitch-switch restart
-
-  if [ $tgt_locks -ne 0 ];then
-    service tgt start
+  if [ $CX == 'ConnectX-4' ]; then
+    service openibd restart && service openvswitch-switch restart
+    #sleep 10
   fi
-
   # Verify VFs
   ./sriov.sh validate
